callautomation-connect-room/requirements.txt
Quart>=0.19.6
python-dotenv==0.21.0
azure-eventgrid==4.11.0
aiohttp>= 3.11.9
azure-communication-callautomation==1.3.0b1
azure-communication-identity==1.3.1
azure-communication-rooms==1.1.1

